requirements.txt
# FastAPI and core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0,<0.38.0
python-multipart>=0.0.9
# mangum>=0.17.0  # Not needed for Render - using Uvicorn directly

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # bcrypt 5.0.0 has compatibility issues with passlib 1.7.4

# Environment and configuration
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator>=2.0.0

# Cloud storage (choose one - boto3 is large, prefer cloudinary for serverless)
# boto3==1.34.0  # Commented out - use cloudinary instead for smaller size
cloudinary==1.36.0

# ML and AI - Using Google Gemini API
google-generativeai>=0.3.0
Pillow>=10.0.0

# Background tasks (not needed for serverless - Vercel handles this)
# celery==5.3.4
# redis==5.0.1

# Email
sendgrid==6.10.0

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2

# Testing (optional - can be removed for production)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2

# Code quality (optional - can be removed for production)
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0

# CORS
fastapi-cors==0.0.6

# PDF (optional - may fail to build on some platforms)
# reportlab>=4.0.0